requirements/base.txt
# Core
pytz==2018.9
Django==2.1.7
django-configurations==2.1

# For the persistence stores
dj-database-url==0.5.0
django-bulk-update==2.2.0


# Rest apis
djangorestframework==3.9.1
django-cors-headers==2.5.0
django-filter==2.1.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.3

# celery
celery==4.2.1
django-celery-beat-1.4.0
numpy==1.16